+++ src/main/webapp/app/entities/feature/feature-util.ts
@@ -30,3 +30,24 @@ export const isFormValid = (feature: IFeature): IError => {
   }
   return error;
 };
+
+export const generateAlternativeDimensionsOptions = data => {
+  const options = [];
+  if (data) {
+    const _data = JSON.parse(data);
+    _data.forEach(function (option) {
+      options.push({ value: option['featureId'], label: option['featureName'] });
+    });
+  }
+  return options;
+};
+
+export const generateFeaturesOptions = data => {
+  const options = [];
+  if (data) {
+    data.forEach(function (option) {
+      options.push({ value: option['id'], label: option['name'] });
+    });
+  }
+  return options;
+};
